requirements.txt
streamlit>=1.28.0
google-generativeai>=0.8.0
PyPDF2==3.0.0
python-docx==1.1.0
protobuf>=4.24.4
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.3
opencv-python-headless>=4.8.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2